REworked the server added main

+++ src/main/java/ru/spbau/mit/Client/Client.java
@@ -7,7 +7,7 @@ import java.io.OutputStream;
 import java.util.List;
 
 public interface Client {
-    void connect(String hostName, int port) throws IOException;
+    void connect(String hostName, short port) throws IOException;
     void disconnect() throws IOException;
     List<RemoteFile> executeList(String path) throws IOException;
     void executeGet(String path, OutputStream out) throws IOException;
